Issue 273 ignore blockTime + changesMinDelta if val === 0
See Issue https://github.com/ioBroker/ioBroker.sql/issues/273

Small quick fix to always log values if value === 0.

Solves problem, that if you log electric power of a solar penal or power of an electric meter of a wallbox, the last value that is logged is often not 0 W. this is because the last value is e.g. 500 W and then 5 seconds later it is 0 W. If you have enabled a blockTime of 20 sec, the last value is alwas 500W. This is confusing in diagrams.

Maybe there's a better solution or this new behavior should be activated with an extra check box?